requirements.txt
matplotlib==3.7.2
numpy==1.25.2
Pillow==10.0.0
scipy==1.11.2
skimage==0.21.0
